+++ CHANGELOG.md
@@ -6,7 +6,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 
 ## 1.1.3 (TBD)
 
-* TBD
+* Point to MGRS conversion latitude min/max bounding and longitude normalization
 
 ## [1.1.2](https://github.com/ngageoint/mgrs-ios/releases/tag/1.1.2) (12-09-2022)